♻️ zb: Move pending method call to separate module
